# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.createPayload:
    properties:
      birthdate:
        type: string
      cpf:
        type: string
      email:
        type: string
      name:
        type: string
    required:
    - birthdate
    - cpf
    - email
    - name
    type: object
  handlers.updatePayload:
    properties:
      birthdate:
        type: string
      cpf:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
  model.User:
    properties:
      birthdate:
        type: string
      cpf:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
  description: This is a Users crud server.
  license:
    name: MIT
  title: Users API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Get all users
      parameters:
      - description: search by name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
      summary: Get all users
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Add user to database
      parameters:
      - description: Add user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.createPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
      summary: Create user
      tags:
      - user
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get a single user by ID
      operationId: string
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      summary: Get a single user by ID
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update user
        in: body
        name: user
        schema:
          $ref: '#/definitions/handlers.updatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      summary: Update user
      tags:
      - user
swagger: "2.0"